Split out all existing loging beyond handling HTTP from the timetable viewing interface, and added this into the domain layer. Existing tests working as expected, but need restructuring
