+++ timetable/domain/view_timetables/__init__.py
@@ -1,0 +1,4 @@
+# Convenience imports
+from .data_pre_processing import (
+    get_summary_stats_for_dashboard, get_year_indexed_pupils, get_letter_indexed_teachers,
+    get_pupil_timetable_context, get_teacher_timetable_context)
